requirements.txt
pikepdf
tabula-py
google-generativeai
flask
firebase-admin